+++ Externals/ffmpeg/include/libavutil/murmur3.h
@@ -27,6 +27,7 @@
 #ifndef AVUTIL_MURMUR3_H
 #define AVUTIL_MURMUR3_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 /**
@@ -97,7 +98,7 @@ void av_murmur3_init(struct AVMurMur3 *c);
  * @param[in]  src  Input data to update hash with
  * @param[in]  len  Number of bytes to read from `src`
  */
-void av_murmur3_update(struct AVMurMur3 *c, const uint8_t *src, int len);
+void av_murmur3_update(struct AVMurMur3 *c, const uint8_t *src, size_t len);
 
 /**
  * Finish hashing and output digest value.
