+++ Externals/ffmpeg/include/libavutil/ripemd.h
@@ -28,10 +28,10 @@
 #ifndef AVUTIL_RIPEMD_H
 #define AVUTIL_RIPEMD_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "attributes.h"
-#include "version.h"
 
 /**
  * @defgroup lavu_ripemd RIPEMD
@@ -66,7 +66,7 @@ int av_ripemd_init(struct AVRIPEMD* context, int bits);
  * @param data    input data to update hash with
  * @param len     input data length
  */
-void av_ripemd_update(struct AVRIPEMD* context, const uint8_t* data, unsigned int len);
+void av_ripemd_update(struct AVRIPEMD* context, const uint8_t* data, size_t len);
 
 /**
  * Finish hashing and output digest value.
